+++ config/config.exs
@@ -2,8 +2,6 @@
 # and its dependencies with the aid of the Mix.Config module.
 use Mix.Config
 
-config :ecto_shorts, repo: nil, error_module: EctoShorts.Actions.Error
-
 if Mix.env() == :test do
   config :ecto_shorts, repo: EctoShorts.Support.TestRepo
 end