+++ lib/ecto_shorts.ex
@@ -1,67 +1,43 @@
 defmodule EctoShorts do
   @moduledoc """
-  Ecto Shorts is a library created to help shorten ecto interactions
-  and remove most of the related code.
+  Ecto Shorts is a library created to help shorten ecto queries
 
-  The major limitation right now is this is restricted to functioning
-  on 1 repo, in the future I would expect multi-repo support to expand
 
-  There are 2 main modules Actions and CommonChanges
+  Common filters available include:
 
-  ### Actions
-  `EctoShorts.Actions` allows for filters to be constructed from data such as
+  - `preload` - Preloads fields onto the query results
+  - `start_date` - Query for items inserted after this date
+  - `end_date` - Query for items inserted before this date
+  - `before` - Get items with IDs before this value
+  - `after` - Get items with IDs after this value
+  - `ids` - Get items with a list of ids
+  - `first` - Gets the first n items
+  - `last` - Gets the last n items
+  - `limit` - Gets the first n items
+  - `offset` - Offsets limit by n items
+  - `search` - ***Warning:*** This requires schemas using this to have a `&by_search(query, val)` function
+  - `join_preload` - WIP: Joins an associations and preloads its fields onto the query results (using only one query)
 
-  ```elixir
-  Actions.all(User, %{
-    favorite_food: "curry",
-    age: %{gte: 18, lte: 50},
-    name: %{ilike: "steven"},
-    preload: [:address],
-    last: 5
-  })
-  ```
+  You are also able to filter on any natural field of a model, as well as use
 
-  which the equivelent would be
-
-  ```elixir
-  Repo.all from u in User,
-    preload: [:address],
-    limit: 5,
-    where: u.favorite_food == "curry" and
-           u.age >= 18 and u.age <= 50 and
-           ilike(u.name, "%steven%")
-  ```
-
-  See `EctoShorts.CommonFilters` for more info on filters
-
-  ### CommonChanges
-  `EctoShorts.CommonChanges` is responsible for updating relations on schemas it allows
-  you to update associations or create using changesets and handles choosing between
-  `put_assoc` and `cast_assoc`
+  - gte/gt
+  - lte/lt
+  - like/ilike
+  - is_nil/not(is_nil)
 
   ```elixir
-  Actions.update(User, 1, %{friends: [%{id: 5, name: "Billy"}]})
-  Actions.update(User, 1, %{friends: [%{id: 5, name: "Bob"}]})
-  ```
+  EctoShorts.filter(User, %{name: %{ilike: "steve"}})
+  EctoShorts.filter(User, %{name: %{age: %{gte: 18, lte: 30}}})
+  EctoShorts.filter(User, %{name: %{is_banned: %{!=: nil}}})
+  EctoShorts.filter(User, %{name: %{is_banned: %{==: nil}}})
 
-  Another feature is you can update many to many relationship members using a id list of members
-
-  ###### How to make user 1 friends with users 5, 6, 7
-  ```elixir
-  Actions.update(User, 1, %{friends: [%{id: 5}, %{id: 6}, %{id: 7}]})
-  ```
-
-  ## Config
-  You can configure both the Repo and Error module via
-  ```elixir
-  config :ecto_shorts,
-    repo: MyRepo,
-    error_module: EctoShorts.Actions.Error
+  my_query = EctoShorts.filter(User, %{name: "Billy"})
+  EctoShorts.filter(my_query, %{last_name: "Joe"})
   ```
   """
 
   def filter(module_or_query, filters, order_by_prop \\ :id) do
-    EctoShorts.CommonFilters.convert_params_to_filter(module_or_query, filters, order_by_prop)
+    EctoShorts.CommonFilters.query_params(module_or_query, filters, order_by_prop)
   end
 
 end