+++ test/actions_test.exs
@@ -1,3 +1,0 @@
-defmodule EctoShorts.ActionTest do
-
-end