+++ test/common_changes_test.exs
@@ -1,3 +1,0 @@
-defmodule EctoShorts.CommonChangesTest do
-
-end